Added Repository, ActivityRepository, RoomRepository, ActivityRepository, PersonRepository + bugfixes

+++ Facultate/Grade.cpp
@@ -25,7 +25,7 @@ void Grade::Print()
 	std::cout << discipline->GetName() << ": " << value << "\n";
 }
 
-int Grade::GetValue()
+float Grade::GetValue()
 {
 	return value;
 }
